src/test/java/com/github/iut/converter/IsPropertyAvailableTagConverterTest.java
package com.github.iut.converter;

import com.github.iut.ext.DocumentExt;
import org.junit.jupiter.api.Test;
import org.w3c.dom.Document;

import static com.github.iut.XmlTestUtils.assertXmlEqual;
import static com.github.iut.XmlTestUtils.loadXML;

class IsPropertyAvailableTagConverterTest {
    @Test
    public void convert1() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source1.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert1.xml");
    }

    @Test
    public void convert2() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source2.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert2.xml");
    }

    @Test
    public void convert3() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source3.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert3.xml");
    }

    @Test
    public void convert4() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source4.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert4.xml");
    }

    @Test
    public void convert5() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source5.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert5.xml");
    }

    @Test
    public void convert6() throws Exception {
        final Document document = loadXML("xmls/IsPropertyAvailableTagConverterTest/source6.xml");
        final Document ret = new IsPropertyAvailableTagConverter().convert(DocumentExt.of(document,"testInput")).getDocument();
        assertXmlEqual(ret, "xmls/IsPropertyAvailableTagConverterTest/expect/convert6.xml");
    }
}